requirements.txt
python-telegram-bot[job-queue]==21.7
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
greenlet==3.1.1
httpx==0.27.2
python-dotenv==1.0.1
aiofiles==24.1.0
